+++ src/Modals/MediaModal.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import styled from 'styled-components';
+
+const Container = styled.div``;
+
+const MediaModal = ({}) => {
+  return(
+    <Container>
+    </Container>
+  );
+};
+
+export default MediaModal;
