+++ src/Modals/index.tsx
@@ -1,0 +1,7 @@
+import Modal from '../Modals/Modal';
+import ConfirmationModal from '../Modals/ConfirmationModal';
+
+export {
+  Modal,
+  ConfirmationModal,
+};
